+++ test-packages/e2e-tests/test/function.spec.ts
@@ -14,7 +14,7 @@ describe('functions', () => {
     const request = returnSapCloudSdk({});
     it('should serialize url', async () => {
       expect(await request.url(destination)).toBe(
-        `${url}odata/test-service/returnSapCloudSdk()?$format=json`
+        `${url}odata/test-service/returnSapCloudSdk()`
       );
     });
 
@@ -39,7 +39,7 @@ describe('functions', () => {
 
     it('should serialize url', async () => {
       expect(await request.url(destination)).toBe(
-        `${url}odata/test-service/concatStrings(Str1='test',Str2='string')?$format=json`
+        `${url}odata/test-service/concatStrings(Str1='test',Str2='string')`
       );
     });